HHH-18693 Changed code to allow creation of metadata for inner static non-private classes           Generaed metadate class for inner class A.B is A_.B_

+++ tooling/metamodel-generator/src/main/java/org/hibernate/processor/HibernateProcessor.java
@@ -7,6 +7,7 @@ package org.hibernate.processor;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.hibernate.processor.annotation.AnnotationMetaEntity;
 import org.hibernate.processor.annotation.AnnotationMetaPackage;
+import org.hibernate.processor.annotation.NonManagedMetamodel;
 import org.hibernate.processor.model.Metamodel;
 import org.hibernate.processor.util.Constants;
 import org.hibernate.processor.xml.JpaDescriptorParser;
@@ -22,6 +23,7 @@ import javax.lang.model.element.AnnotationValue;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.ExecutableElement;
+import javax.lang.model.element.Modifier;
 import javax.lang.model.element.PackageElement;
 import javax.lang.model.element.QualifiedNameable;
 import javax.lang.model.element.TypeElement;
@@ -82,6 +84,7 @@ import static org.hibernate.processor.util.TypeUtils.getAnnotationMirror;
 import static org.hibernate.processor.util.TypeUtils.getAnnotationValue;
 import static org.hibernate.processor.util.TypeUtils.hasAnnotation;
 import static org.hibernate.processor.util.TypeUtils.isClassOrRecordType;
+import static org.hibernate.processor.util.TypeUtils.isMemberType;
 
 /**
  * Main annotation processor.
@@ -358,57 +361,99 @@ public class HibernateProcessor extends AbstractProcessor {
 		}
 
 		for ( Element element : roundEnvironment.getRootElements() ) {
-			try {
-				if ( !included( element )
-						|| hasAnnotation( element, Constants.EXCLUDE )
-						|| hasPackageAnnotation( element, Constants.EXCLUDE ) ) {
-					// skip it completely
+			processElement( element, null );
+		}
+	}
+
+	private void processElement(Element element, @Nullable Element parent) {
+		try {
+			if ( !included( element )
+					|| hasAnnotation( element, Constants.EXCLUDE )
+					|| hasPackageAnnotation( element, Constants.EXCLUDE )
+					|| element.getModifiers().contains( Modifier.PRIVATE ) ) {
+				// skip it completely
+				return;
+			}
+			else if ( isEntityOrEmbeddable( element ) && !element.getModifiers().contains( Modifier.PRIVATE )) {
+				context.logMessage( Diagnostic.Kind.OTHER, "Processing annotated entity class '" + element + "'" );
+				handleRootElementAnnotationMirrors( element, parent );
+			}
+			else if ( hasAuxiliaryAnnotations( element ) ) {
+				context.logMessage( Diagnostic.Kind.OTHER, "Processing annotated class '" + element + "'" );
+				handleRootElementAuxiliaryAnnotationMirrors( element );
+			}
+			else if ( element instanceof TypeElement typeElement ) {
+				final AnnotationMirror repository = getAnnotationMirror( element, JD_REPOSITORY );
+				if ( repository != null ) {
+					final AnnotationValue provider = getAnnotationValue( repository, "provider" );
+					if ( provider == null
+							|| provider.getValue().toString().isEmpty()
+							|| provider.getValue().toString().equalsIgnoreCase("hibernate") ) {
+						context.logMessage( Diagnostic.Kind.OTHER, "Processing repository class '" + element + "'" );
+						final AnnotationMetaEntity metaEntity =
+								AnnotationMetaEntity.create( typeElement, context );
+						if ( metaEntity.isInitialized() ) {
+							context.addMetaAuxiliary( metaEntity.getQualifiedName(), metaEntity );
+						}
+						// otherwise discard it (assume it has query by magical method name stuff)
+					}
 				}
-				else if ( isEntityOrEmbeddable( element ) ) {
-					context.logMessage( Diagnostic.Kind.OTHER, "Processing annotated entity class '" + element + "'" );
-					handleRootElementAnnotationMirrors( element );
-				}
-				else if ( hasAuxiliaryAnnotations( element ) ) {
-					context.logMessage( Diagnostic.Kind.OTHER, "Processing annotated class '" + element + "'" );
-					handleRootElementAuxiliaryAnnotationMirrors( element );
-				}
-				else if ( element instanceof TypeElement typeElement ) {
-					final AnnotationMirror repository = getAnnotationMirror( element, JD_REPOSITORY );
-					if ( repository != null ) {
-						final AnnotationValue provider = getAnnotationValue( repository, "provider" );
-						if ( provider == null
-								|| provider.getValue().toString().isEmpty()
-								|| provider.getValue().toString().equalsIgnoreCase("hibernate") ) {
-							context.logMessage( Diagnostic.Kind.OTHER, "Processing repository class '" + element + "'" );
+				else {
+					for ( Element member : typeElement.getEnclosedElements() ) {
+						if ( hasAnnotation( member, HQL, SQL, FIND ) ) {
+							context.logMessage( Diagnostic.Kind.OTHER, "Processing annotated class '" + element + "'" );
 							final AnnotationMetaEntity metaEntity =
 									AnnotationMetaEntity.create( typeElement, context );
-							if ( metaEntity.isInitialized() ) {
-								context.addMetaAuxiliary( metaEntity.getQualifiedName(), metaEntity );
-							}
-							// otherwise discard it (assume it has query by magical method name stuff)
+							context.addMetaAuxiliary( metaEntity.getQualifiedName(), metaEntity );
+							break;
 						}
 					}
-					else {
-						for ( Element member : typeElement.getEnclosedElements() ) {
-							if ( hasAnnotation( member, HQL, SQL, FIND ) ) {
-								context.logMessage( Diagnostic.Kind.OTHER, "Processing annotated class '" + element + "'" );
-								final AnnotationMetaEntity metaEntity =
-										AnnotationMetaEntity.create( typeElement, context );
-								context.addMetaAuxiliary( metaEntity.getQualifiedName(), metaEntity );
-								break;
+					if ( enclosesEntityOrEmbeddable( element ) ) {
+						AnnotationMetaEntity parentMeta = null;
+						if ( parent instanceof TypeElement parentElement ) {
+							final String key = parentElement.getQualifiedName().toString();
+							if ( context.getMetamodel( key ) instanceof AnnotationMetaEntity parentMetaEntity ) {
+								parentMeta = parentMetaEntity;
 							}
 						}
+						final NonManagedMetamodel  metaEntity =
+								NonManagedMetamodel .create(
+										typeElement, context,
+										false, parentMeta );
+						context.addMetaEntity( metaEntity.getQualifiedName(), metaEntity );
+						if ( context.generateJakartaDataStaticMetamodel()) {
+							AnnotationMetaEntity parentDataMeta = null;
+							if ( parent instanceof TypeElement parentElement ) {
+								final String key = parentElement.getQualifiedName().toString();
+								if ( context.getDataMetaEntity( key ) instanceof AnnotationMetaEntity parentMetaEntity ) {
+									parentDataMeta = parentMetaEntity;
+								}
+							}
+							final NonManagedMetamodel  dataMetaEntity =
+									NonManagedMetamodel .create(
+											typeElement, context,
+											true, parentDataMeta );
+							context.addDataMetaEntity( dataMetaEntity.getQualifiedName(), dataMetaEntity );
+						}
+
 					}
 				}
 			}
-			catch ( ProcessLaterException processLaterException ) {
-				if ( element instanceof TypeElement ) {
-					context.logMessage(
-							Diagnostic.Kind.OTHER,
-							"Could not process '" + element + "' (will redo in next round)"
-					);
-					context.addElementToRedo( ( (TypeElement) element).getQualifiedName() );
+			if ( isClassOrRecordType( element ) ) {
+				for ( final Element child : element.getEnclosedElements() ) {
+					if ( isClassOrRecordType( child ) ) {
+						processElement( child, element );
+					}
 				}
+			}
+		}
+		catch ( ProcessLaterException processLaterException ) {
+			if ( element instanceof TypeElement ) {
+				context.logMessage(
+						Diagnostic.Kind.OTHER,
+						"Could not process '" + element + "' (will redo in next round)"
+				);
+				context.addElementToRedo( ( (TypeElement) element ).getQualifiedName() );
 			}
 		}
 	}
@@ -430,7 +475,7 @@ public class HibernateProcessor extends AbstractProcessor {
 		}
 
 		for ( Metamodel entity : context.getMetaEntities() ) {
-			if ( !context.isAlreadyGenerated(entity) ) {
+			if ( !context.isAlreadyGenerated( entity ) && !isMemberType( entity.getElement() ) ) {
 				context.logMessage( Diagnostic.Kind.OTHER,
 						"Writing Jakarta Persistence metamodel for entity '" + entity + "'" );
 				ClassWriter.writeFile( entity, context );
@@ -439,7 +484,7 @@ public class HibernateProcessor extends AbstractProcessor {
 		}
 
 		for ( Metamodel entity : context.getDataMetaEntities() ) {
-			if ( !context.isAlreadyGenerated(entity) ) {
+			if ( !context.isAlreadyGenerated( entity ) && !isMemberType( entity.getElement() ) ) {
 				context.logMessage( Diagnostic.Kind.OTHER,
 						"Writing Jakarta Data metamodel for entity '" + entity + "'" );
 				ClassWriter.writeFile( entity, context );
@@ -516,6 +561,18 @@ public class HibernateProcessor extends AbstractProcessor {
 		return false;
 	}
 
+	private static boolean enclosesEntityOrEmbeddable(Element element) {
+		if ( !(element instanceof TypeElement typeElement) ) {
+			return false;
+		}
+		for ( final Element enclosedElement : typeElement.getEnclosedElements() ) {
+			if ( isEntityOrEmbeddable( enclosedElement ) || enclosesEntityOrEmbeddable( enclosedElement ) ) {
+				return true;
+			}
+		}
+		return false;
+	}
+
 	private static boolean isEntityOrEmbeddable(Element element) {
 		return hasAnnotation(
 				element,
@@ -547,7 +604,7 @@ public class HibernateProcessor extends AbstractProcessor {
 		);
 	}
 
-	private void handleRootElementAnnotationMirrors(final Element element) {
+	private void handleRootElementAnnotationMirrors(final Element element, @Nullable Element parent) {
 		if ( isClassOrRecordType( element ) ) {
 			if ( isEntityOrEmbeddable( element ) ) {
 				final TypeElement typeElement = (TypeElement) element;
@@ -564,12 +621,20 @@ public class HibernateProcessor extends AbstractProcessor {
 									+ "' since XML configuration is metadata complete.");
 				}
 				else {
+					AnnotationMetaEntity parentMetaEntity = null;
+					if ( parent instanceof TypeElement parentTypeElement ) {
+						if ( context.getMetamodel(
+								parentTypeElement.getQualifiedName().toString() )
+								instanceof AnnotationMetaEntity pme ) {
+							parentMetaEntity = pme;
+						}
+					}
 					final boolean requiresLazyMemberInitialization
 							= hasAnnotation( element, EMBEDDABLE, MAPPED_SUPERCLASS );
 					final AnnotationMetaEntity metaEntity =
 							AnnotationMetaEntity.create( typeElement, context,
 									requiresLazyMemberInitialization,
-									true, false );
+									true, false, parentMetaEntity );
 					if ( alreadyExistingMetaEntity != null ) {
 						metaEntity.mergeInMembers( alreadyExistingMetaEntity );
 					}
@@ -583,10 +648,17 @@ public class HibernateProcessor extends AbstractProcessor {
 							// let a handwritten metamodel "override" the generated one
 							// (this is used in the Jakarta Data TCK)
 							&& !hasHandwrittenMetamodel(element) ) {
+						AnnotationMetaEntity parentDataEntity = null;
+						if ( parent instanceof TypeElement parentTypeElement ) {
+							if ( context.getDataMetaEntity( parentTypeElement.getQualifiedName().toString() )
+									instanceof AnnotationMetaEntity pme ) {
+								parentDataEntity = pme;
+							}
+						}
 						final AnnotationMetaEntity dataMetaEntity =
 								AnnotationMetaEntity.create( typeElement, context,
 										requiresLazyMemberInitialization,
-										true, true );
+										true, true, parentDataEntity );
 //						final Metamodel alreadyExistingDataMetaEntity =
 //								tryGettingExistingDataEntityFromContext( mirror, '_' + qualifiedName );
 //						if ( alreadyExistingDataMetaEntity != null ) {
